chore: move logger out of root
