+++ packages/strapi-plugin-status-manager/server/src/logger.ts
@@ -2,4 +2,4 @@ import { DebugLogger } from "@primer/debugger";
 
 const logger = new DebugLogger("StatusManager");
 
-export default logger;
+export { logger };
